Get resfile.c building successfully

+++ app/src/main/cpp/Libraries/RES/resbuild.c
@@ -47,7 +47,7 @@ bool ResEraseIfInFile(Id id);                // erase item from file
 //    For Mac version:  Does nothing.  May go back later and add comment via the
 // desktop database, maybe.
 
-void ResSetComment(int16_t filenum, const char *comment)
+void ResSetComment(int32_t filenum, const char *comment)
 {
     ResFileHeader *phead;
